Fix a bug that prevented paging of help when 'help' option was not defined manually.
'page_help' parameter result with the addition (in the 'ACTIONS' variable) of a
custom 'HelpPager' action for replacing the default action ('_HelpAction') of
the 'help' option. But actions are registered after the parser is initialized
so, if 'add_help' parameter was not set (or set to 'True'), the 'help' option
was added using the default 'argparse._HelpAction' action.

Now 'argparse._HelpAction' is replaced by the 'HelpPager' action if 'page_help'
is set.