Improve output the 'help' command.
The output of the 'help' command (added by the 'add_help_cmd' parameter)
is now more like the tree command (with three symbols and a space before
an entry).
The calculation of the index to start printing the description has also
been improved and made more reliable (it starts four spaces after the
longest command).